+++ setup/ReplaceFonts.jsx
@@ -1,5 +1,5 @@
 ﻿/*
-	Replace fonts 2.2 (2021-10-11)
+	Replace fonts 2.2.1 (2021-10-11)
 	(c) 2020-2021 Paul Chiorean (jpeg@basement.ro)
 
 	Replaces fonts from a 4-column TSV file named 'fonts.txt':
@@ -47,7 +47,7 @@ function main() {
 	var VERBOSITY = 1; // 0: FAIL, 1: +WARN, 2: +INFO
 	var file, data, messages;
 	var counter = 0;
-	if (!(file = getDataFile('fonts.txt'))) { alert('No data file found.'); exit(); }
+	if (!(file = getDataFile('fonts.txt'))) { if (VERBOSITY > 1) alert('No data file found.'); exit(); }
 	data = parseDataFile(file);
 	if (data.errors.fail.length > 0) { report(data.errors.fail, decodeURI(file.getRelativeURI(doc.filePath))); exit(); }
 	if (data.records.length === 0) exit();